Fixed signed/unsigned conversions (XERCESC-1531)
git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@557254 13f79535-47bb-0310-9956-ffa450edef68

+++ src/xercesc/util/HexBin.cpp
@@ -102,7 +102,7 @@ XMLByte* HexBin::decodeToXMLByte(const XMLCh*          const   hexData
         return 0;
 
     //prepare the return string
-    int decodeLength = strLen/2;
+    int decodeLength = (int)strLen/2;
     XMLByte *retVal = (XMLByte*) manager->allocate( (decodeLength + 1) * sizeof(XMLByte));
     ArrayJanitor<XMLByte> janFill(retVal, manager);
